backend/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.7.1
django-cors-headers==3.11.0
django-colorfield
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==1.7.0
pytz==2020.1
sqlparse==0.3.1 
python-dotenv==0.10.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
Pillow==9.1.1
django-rest-auth==0.9.5
django-allauth==0.51.0
djoser==2.1.0
drf-extra-fields==3.2.1